task-40: Add workspace editing

+++ client/src/store/modules/workspaces/actions.ts
@@ -22,7 +22,7 @@ export const actions: ActionTree<WorkspacesState, State> = {
 	async create({ commit, rootGetters }, payload: { name: string }): Promise<void> {
 		try {
 			const { data } = await WorkspaceApi.create({ name: payload.name }, rootGetters['user/accessToken'])
-			// TODO: commit('addWorkspace', { workspace: data.workspace })
+			// TODO: commit('add', { workspace: data.workspace })
 		} catch (error) {
 			if (axios.isAxiosError(error)) {
 				if (error.response?.status === 401) {
@@ -34,17 +34,44 @@ export const actions: ActionTree<WorkspacesState, State> = {
 			}
 		}
 	},
-	async delete({ commit, rootGetters }, payload: { id: number }): Promise<void> {
-		try {
-			await WorkspaceApi.delete({ id: payload.id }, rootGetters['user/accessToken'])
-			commit('deleteWorkspace', { id: payload.id })
-		} catch (error) {
-			if (axios.isAxiosError(error)) {
-				if (error.response?.status === 401) {
-					// TODO: 401 handler
-					throw new Error('Access token is incorrect')
+	async delete({ commit, state, rootGetters }, payload: { id: number }): Promise<void> {
+		const workspaceIndex = state.workspaces.findIndex(item => item.id === payload.id);
+		if (~workspaceIndex) {
+			try {
+				await WorkspaceApi.delete({ id: payload.id }, rootGetters['user/accessToken'])
+				commit('delete', { workspaceIndex: workspaceIndex })
+			} catch (error) {
+				if (axios.isAxiosError(error)) {
+					if (error.response?.status === 401) {
+						// TODO: 401 handler
+						throw new Error('Access token is incorrect')
+					} else {
+						throw new Error(error.response?.statusText);
+					}
 				}
 			}
+		} else {
+			throw new Error('Workspace id is incorrect')
+		}
+	},
+	async update({ commit, state, rootGetters }, payload: { id: number, name: string }): Promise<void> {
+		const workspaceIndex = state.workspaces.findIndex(item => item.id === payload.id);
+		if (~workspaceIndex) {
+			try {
+				const { data } = await WorkspaceApi.update({ id: payload.id, name: payload.name }, rootGetters['user/accessToken'])
+				// commit('update', { oldWorkspaceIndex: workspaceIndex, newWorkspace: data.workspace });
+			} catch (error) {
+				if(axios.isAxiosError(error)) {
+					if (error.response?.status === 401) {
+						// TODO: 401 handler
+						throw new Error('Access token is incorrect')
+					} else {
+						throw new Error(error.response?.statusText);
+					}
+				}
+			}
+		} else {
+			throw new Error('Workspace id is incorrect')
 		}
 	}
 }
